Refactor interactive chat and add token counter test
